Create account
Perhaps the simplest software that you can work with is the one that allows you to deal with bank accounts and transactions regarding it. Designing a robust system that allows you to engage in transactions is something that every beginner should get started with. The proposed system is a web-based project that allows you to do everything a bank would allow you to do naturally.

One should be able to deposit money and withdraw money from a particular account as the user desires. There should be a validation to allow only a particular amount of cash inflows at any time, as well as to allow withdraws if the balance is sufficient. There should also be the calculation of interest and its addition to the balance every month.

There can be multiple improvements for this project, including adding support for multiple types of accounts. For example, there can be a recurring deposit account that should not allow withdrawals till the maturation date, and the amount is automatically added to the account post the maturation date. We started with this because it is still one of the most popular java projects for beginners.